# [k8s] Optimize `/kubernetes_node_info` for clusters with no accelerators by 3x