--- a/tests/unit_tests/kubernetes/test_kubernetes_utils.py
+++ b/tests/unit_tests/kubernetes/test_kubernetes_utils.py
@@ -32,23 +32,23 @@ def test_get_kubernetes_nodes():
 def test_get_kubernetes_node_info():
     """Tests get_kubernetes_node_info function."""
     # Mock node and pod objects
-    mock_node_1 = mock.MagicMock()
-    mock_node_1.metadata.name = 'node-1'
-    mock_node_1.metadata.labels = {
+    mock_gpu_node_1 = mock.MagicMock()
+    mock_gpu_node_1.metadata.name = 'node-1'
+    mock_gpu_node_1.metadata.labels = {
         'skypilot.co/accelerator': 'a100-80gb',
         'cloud.google.com/gke-accelerator-count': '4'
     }
-    mock_node_1.status.allocatable = {'nvidia.com/gpu': '4'}
+    mock_gpu_node_1.status.allocatable = {'nvidia.com/gpu': '4'}
 
-    mock_node_2 = mock.MagicMock()
-    mock_node_2.metadata.name = 'node-2'
-    mock_node_2.metadata.labels = {
+    mock_gpu_node_2 = mock.MagicMock()
+    mock_gpu_node_2.metadata.name = 'node-2'
+    mock_gpu_node_2.metadata.labels = {
         'skypilot.co/accelerator': 'tpu-v4-podslice',
         'cloud.google.com/gke-accelerator-count': '8',
         'cloud.google.com/gke-tpu-accelerator': 'tpu-v4-podslice',
         'cloud.google.com/gke-tpu-topology': '2x4'
     }
-    mock_node_2.status.allocatable = {'google.com/tpu': '8'}
+    mock_gpu_node_2.status.allocatable = {'google.com/tpu': '8'}
 
     mock_pod_1 = mock.MagicMock()
     mock_pod_1.spec.node_name = 'node-1'
@@ -68,7 +68,7 @@ def test_get_kubernetes_node_info():
 
     # Test case 1: Normal operation with GPU and TPU nodes
     with mock.patch('sky.provision.kubernetes.utils.get_kubernetes_nodes',
-                   return_value=[mock_node_1, mock_node_2]), \
+                   return_value=[mock_gpu_node_1, mock_gpu_node_2]), \
          mock.patch('sky.provision.kubernetes.utils.'
                    'get_all_pods_in_kubernetes_cluster',
                    return_value=[mock_pod_1, mock_pod_2]), \
@@ -92,7 +92,7 @@ def test_get_kubernetes_node_info():
 
     # Test case 2: No permission to list pods
     with mock.patch('sky.provision.kubernetes.utils.get_kubernetes_nodes',
-                   return_value=[mock_node_1, mock_node_2]), \
+                   return_value=[mock_gpu_node_1, mock_gpu_node_2]), \
          mock.patch('sky.provision.kubernetes.utils.'
                    'get_all_pods_in_kubernetes_cluster',
                    side_effect=utils.kubernetes.kubernetes.client.ApiException(
@@ -118,7 +118,7 @@ def test_get_kubernetes_node_info():
     mock_node_3.status.allocatable = {'google.com/tpu': '4'}
 
     with mock.patch('sky.provision.kubernetes.utils.get_kubernetes_nodes',
-                   return_value=[mock_node_1, mock_node_3]), \
+                   return_value=[mock_gpu_node_1, mock_node_3]), \
          mock.patch('sky.provision.kubernetes.utils.'
                    'get_all_pods_in_kubernetes_cluster',
                    return_value=[mock_pod_1]):
@@ -139,6 +139,65 @@ def test_get_kubernetes_node_info():
         assert isinstance(node_info, models.KubernetesNodesInfo)
         assert len(node_info.node_info_dict) == 0
         assert node_info.hint == ''
+
+    # Test case 5: CPU-only nodes
+    mock_cpu_node_1 = mock.MagicMock()
+    mock_cpu_node_1.metadata.name = 'cpu-node-1'
+    mock_cpu_node_1.metadata.labels = {}
+    mock_cpu_node_1.status.allocatable = {'cpu': '4', 'memory': '16Gi'}
+    mock_cpu_node_1.status.addresses = [
+        mock.MagicMock(type='InternalIP', address='10.0.0.1')
+    ]
+
+    mock_cpu_node_2 = mock.MagicMock()
+    mock_cpu_node_2.metadata.name = 'cpu-node-2'
+    mock_cpu_node_2.metadata.labels = {}
+    mock_cpu_node_2.status.allocatable = {'cpu': '8', 'memory': '32Gi'}
+    mock_cpu_node_2.status.addresses = [
+        mock.MagicMock(type='InternalIP', address='10.0.0.2')
+    ]
+
+    with mock.patch('sky.provision.kubernetes.utils.get_kubernetes_nodes',
+                   return_value=[mock_cpu_node_1, mock_cpu_node_2]), \
+         mock.patch('sky.provision.kubernetes.utils.'
+                   'get_all_pods_in_kubernetes_cluster') as mock_get_pods:
+        node_info = utils.get_kubernetes_node_info()
+
+        mock_get_pods.assert_not_called()
+        assert isinstance(node_info, models.KubernetesNodesInfo)
+        assert len(node_info.node_info_dict) == 2
+        assert node_info.node_info_dict['cpu-node-1'].accelerator_type is None
+        assert node_info.node_info_dict['cpu-node-1'].total[
+            'accelerator_count'] == 0
+        assert node_info.node_info_dict['cpu-node-1'].free[
+            'accelerators_available'] == 0
+        assert node_info.node_info_dict['cpu-node-2'].total[
+            'accelerator_count'] == 0
+        assert node_info.node_info_dict['cpu-node-2'].free[
+            'accelerators_available'] == 0
+
+    # Test case 6: Mixed CPU and GPU nodes
+    with mock.patch('sky.provision.kubernetes.utils.get_kubernetes_nodes',
+                   return_value=[mock_cpu_node_1, mock_gpu_node_1]), \
+         mock.patch('sky.provision.kubernetes.utils.'
+                   'get_all_pods_in_kubernetes_cluster',
+                   return_value=[mock_pod_1]) as mock_get_pods, \
+         mock.patch('sky.provision.kubernetes.utils.get_gpu_resource_key',
+                   return_value='nvidia.com/gpu'):
+        node_info = utils.get_kubernetes_node_info()
+
+        mock_get_pods.assert_called_once()
+        assert len(node_info.node_info_dict) == 2
+        # CPU node should have 0 accelerators
+        assert node_info.node_info_dict['cpu-node-1'].total[
+            'accelerator_count'] == 0
+        assert node_info.node_info_dict['cpu-node-1'].free[
+            'accelerators_available'] == 0
+        # GPU node should have correct allocation
+        assert node_info.node_info_dict['node-1'].total[
+            'accelerator_count'] == 4
+        assert node_info.node_info_dict['node-1'].free[
+            'accelerators_available'] == 2
 
 
 def test_get_all_kube_context_names():