In `get_kubernetes_node_info`, we call `get_all_pods_in_kubernetes_cluster`, which is expensive for clusters with a lot of pods. The reason we need to call this is to get the real-time resource usage of accelerators, which we get by reading each pod's `pod.spec.containers[*].resources.requests`.

But this is only relevant for clusters with accelerators, we can skip this otherwise. This PR optimizes for this case by moving the call to `get_all_pods_in_kubernetes_cluster` down, after we check that at least one node have `accelerator_count` > 0.

Tested on a CPU-only cluster with 1000 pods and 100 requests each to `/kubernetes_node_info`:
- Before: 
```
Min:  2836.28ms
p50:  2947.82ms
p95:  3849.55ms
p99:  7867.67ms
Max:  7896.89ms
```
- After: 
```
Min:  611.65ms
p50:  717.43ms
p95:  1115.72ms
p99:  1421.59ms
Max:  1423.58ms
```

Future note: There might also be room for optimization for heterogenous clusters with `number_of_cpu_nodes` >>> `number_of_gpu_nodes`, where instead of getting all pods, we can get the pods only in GPU nodes, using field or label selectors. But this needs more benchmarking and is unclear whether it's worth it.

Tested (run the relevant ones):

- [x] Code formatting: install pre-commit (auto-check on commit) or `bash format.sh`
- [x] Any manual or new tests for this PR (please specify below)
- [x] Backward compatibility: `/quicktest-core` (CI) or `pytest tests/smoke_tests/test_backward_compat.py` (local)

<!-- CI commands (/-prefixed) can only be triggered by repo members -->