@SeungjinYang 2025-10-01
For further optimization, we can probably also organize the code to be:
1. figure out which node has accelerators.
2. organize pods into nodes. Disregard any pods that belong to nodes without accelerators, or any pods that doesn't have accelerators themselves.
3. ...