@kevinmingtarja 2025-10-01
> For further optimization, we can probably also organize the code to be:
> 
> 1. figure out which node has accelerators.
> 2. organize pods into nodes. Disregard any pods that belong to nodes without accelerators, or any pods that doesn't have accelerators themselves.
> 3. ...

Tried a bunch of things, and TL;DR, we can't filter pods by multiple nodes just using a single call, as [field selectors](https://kubernetes.io/docs/concepts/overview/working-with-objects/field-selectors/) does not support it, so we'd need to do N_GPU calls. But I managed to make it such that we precompute the number of allocated accelerators per node, so the for loop over the nodes just need to do a single map lookup.